stories_microservices/auth_service/requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.0
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-slugify==6.1.2
six==1.16.0
SQLAlchemy==1.4.39
text-unidecode==1.3
Werkzeug==2.2.1
zipp==3.8.1
